+++ src/layouts/layout_comparison_page/layout_comparison_page_tab.py
@@ -18,12 +18,12 @@ columns_defs_1 = [
      "initialWidth": 200},
 
     {"field": "run_name",
-     "headerName": "Run naam",
-     "editable": False,
+     "headerName": "Naam berekening",
+     "editable": True,
      "initialWidth": 200},
 
     {"field": "active",
-     "headerName": "Aktieveer",
+     "headerName": "Activeer",
      "cellRenderer": "DBC_Switch",
      "editable": True,
      "CellRendererParams": {"onColor": "success", "offColor": "danger"},
@@ -38,8 +38,10 @@ table_imported_dike_data = dag.AgGrid(
                    "wrapHeaderText": True,
                    "autoHeaderHeight": True, },
     dashGridOptions={"rowSelection": "multiple", "enableCellTextSelection": True, "ensureDomOrder": True},
-    persistence=True,
-    persistence_type="session",
+
+
+    # persistence=True,
+    # persistence_type="session",
 
 )
 
@@ -48,12 +50,12 @@ table_imported_dike_data = dag.AgGrid(
 
 
 left_side = [
-
+    #add some explanation text
     dcc.Upload(
         id='upload-dike-data-comparison',
         children=html.Div([
             '',
-            html.A('Selecteer een bestand dike_data.json')
+            html.A('Selecteer een json-bestand van een traject')
         ]),
         style={
             'width': '100%',
@@ -73,7 +75,7 @@ left_side = [
     dbc.Accordion([
         dbc.AccordionItem(
             [table_imported_dike_data],
-            title='Geimporteerde runs',
+            title='Geimporteerde berekeningen',
         ),
     ]),
 
@@ -84,11 +86,11 @@ right_side = [
     dbc.Tabs(
         [
             dbc.Tab(label="Kaart", tab_id="tab-11111"),
-            dbc.Tab(label="Kosten", tab_id="tab-11112"),
-            dbc.Tab(label="Tijd", tab_id="tab-11113"),
-            dbc.Tab(label="Maatregelen", tab_id="tab-11114"),
-            dbc.Tab(label="Tafel", tab_id="tab-11115"),
-            dbc.Tab(label="PLACEHOLDER 2", tab_id="tab-11116"),
+            dbc.Tab(label="Resultaten optimalisatie", tab_id="tab-11112"),
+            dbc.Tab(label="Faalkans in tijd", tab_id="tab-11113"),
+            dbc.Tab(label="Maatregelen op kaart", tab_id="tab-11114"),
+            dbc.Tab(label="Overzichtstabel maatregelen", tab_id="tab-11115"),
+            dbc.Tab(label="Sectie volgorde", tab_id="tab-11116"),
         ],
         id=TABS_SWITCH_VISUALIZATION_COMPARISON_PAGE,
         active_tab="tab-11111",  # Set the initial active tab