+++ src/layouts/layout_comparison_page/layout_output_tabs.py
@@ -4,6 +4,7 @@ import dash_bootstrap_components as dbc
 from src.component_ids import OVERVIEW_COMPARISON_MAP_ID, RUNS_COMPARISON_GRAPH_ID, RUNS_COMPARISON_GRAPH_TIME_ID, \
     MEASURE_COMPARISON_MAP_ID
 from src.layouts.layout_comparison_page.layout_table_comparison import table_ag_grid_comparison_measures
+from src.layouts.layout_comparison_page.layout_table_comparison_order_section import table_ag_grid_order_measures
 from src.layouts.layout_traject_page.layout_radio_items import layout_radio_result_type_comparison_page
 from src.layouts.layout_traject_page.layout_sliders import layout_year_slider
 from src.plotly_graphs.pf_length_cost import plot_default_scatter_dummy
@@ -16,7 +17,7 @@ def layout_project_output_tab_one() -> html.Div:
     return html.Div(
         children=[
             dbc.Row([
-                dbc.Col([html.H2("Overzicht Projects")], md=10),
+                dbc.Col([html.H2("Overzicht dijktrajecten")], md=10),
             ]),
 
             html.Div(id=OVERVIEW_COMPARISON_MAP_ID,
@@ -29,7 +30,7 @@ def layout_project_output_tab_two() -> html.Div:
     return html.Div(
         children=[
             dbc.Row([
-                dbc.Col([html.H2("Vergelijken kosten")], md=10),
+                dbc.Col([html.H2("Vergelijking kosten en faalkans")], md=10),
             ]),
             layout_year_slider,
             html.Div(
@@ -49,7 +50,7 @@ def layout_project_output_tab_three() -> html.Div:
     return html.Div(
         children=[
             dbc.Row([
-                dbc.Col([html.H2("Tijd ")], md=10),
+                dbc.Col([html.H2("Faalkans in de tijd")], md=10),
             ]),
 
             html.Div(
@@ -88,4 +89,12 @@ def layout_project_output_tab_five() -> html.Div:
         children=[
 
             table_ag_grid_comparison_measures,
+        ])
+
+
+def layout_project_output_tab_six() -> html.Div:
+    return html.Div(
+        children=[
+
+            table_ag_grid_order_measures,
         ])